pyproject.toml
[project]
name = "pulsar-api"
version = "0.3.0"
description = "FastAPI backend with RAG pipeline and LLM streaming for the Pulsar learning project"
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "uvicorn[standard]>=0.42.0",
    "anthropic>=0.85.0",
    "llama-index-core>=0.14.18",
    "llama-index-readers-file>=0.6.0",
    "llama-index-embeddings-huggingface>=0.7.0",
    "chromadb>=1.5.5",
    "llama-index-vector-stores-chroma>=0.5.5",
    "pydantic-settings>=2.13.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.28.1",
    "ruff>=0.15.6",
    "mypy>=1.19.1",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "ANN", "B", "A", "SIM"]
ignore = []

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
exclude = ['\\.venv']
